Add MingW and Unix CMake builds
Both are disabled by default. MingW because we probably
only want to do them on one specific worker. CMake because
we will need to verify individual workers before enabling
the builds. Since the Unix builds are not officially
"supported" we might not care to limit them to a smaller
number of platforms.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>