jenkins/windows-server: Drop Windows Server 2019 variants
We haven't used them in quite some time, so let's not
continue to waste time building them.

Does not touch buildbot-host/buildbot-worker-windows-server-2019.
While that is even more unused (and probably broken), it might
still be useful as a starting point if we want to start running
our own Windows build agents again.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>