buildmaster: Fix mail notifiers
Having the mail notifications done by owners did
not work as we wanted it to since it required us
to enable notifying involved persons. Which was
just a good way to spam people that didn't expect
it. We want the mails to *only* go to the mailing
lists (and some admins).

So instead now we create separate mail notifiers
for each project and use extraRecipients to send
the mails.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>